Move stocktake lines update to respective folder/files
